汇总接口增加排序

+++ obd-core/src/main/java/com/wkhmedical/dto/CarInfoDTO.java
@@ -92,11 +92,11 @@ public class CarInfoDTO implements Serializable {
 	private String maintUrTel;
 
 	@ApiModelProperty(value = "省ID")
-	private Integer provId;
+	private Long provId;
 
 	@ApiModelProperty(value = "市ID")
-	private Integer cityId;
+	private Long cityId;
 
 	@ApiModelProperty(value = "区县ID")
-	private Integer areaId;
+	private Long areaId;
 }
